requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
psycopg2-binary>=2.9.9
supabase>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
